+++ src/main/java/utils/EnumMapUtils.java
@@ -1,13 +1,11 @@
 package utils;
 
-import model.HttpHeaders;
-
 import java.util.EnumMap;
 import java.util.Map;
 
 public class EnumMapUtils {
 
-    public static <K extends Enum<K>,V> EnumMap<K , V> of(Map<K,V> map) {
-        return new EnumMap<K,V>(map);
+    public static <K extends Enum<K>, V> EnumMap<K, V> of(Map<K, V> map) {
+        return new EnumMap<K, V>(map);
     }
 }
