+++ src/main/java/webserver/http/HttpHeaders.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import java.util.Arrays;
