+++ src/main/java/webserver/http/HttpMethod.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import java.util.Arrays;
 import java.util.Optional;
