+++ src/main/java/webserver/http/HttpRequestBody.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import utils.RequestUtils;
