+++ src/main/java/webserver/http/ResponseHeader.java
@@ -1,8 +1,7 @@
-package model;
+package webserver.http;
 
 import java.util.Collections;
 import java.util.EnumMap;
-import java.util.HashMap;
 import java.util.Map;
 
 public class ResponseHeader {
