+++ src/main/java/webserver/http/exception/HttpMethodExceptionMessage.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.http.exception;
 
 public class HttpMethodExceptionMessage {
     public static final String NOT_FOUND_MESSAGE = "HTTP METHOD 를 찾을 수 없습니다.";
