+++ src/main/java/webserver/http/exception/NotFoundHttpMethodException.java
@@ -1,4 +1,4 @@
-package exception;
+package webserver.http.exception;
 
 public class NotFoundHttpMethodException extends RuntimeException {
     public NotFoundHttpMethodException(String message) {
