+++ src/test/java/controller/CreateUserControllerTest.java
@@ -1,9 +1,9 @@
 package controller;
 
-import model.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.http.*;
 
 import java.util.List;
 import java.util.Map;
