+++ src/test/java/webserver/http/HttpRequestHeaderTest.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @DisplayName("HttpRequestHeader 테스트")
 class HttpRequestHeaderTest {
@@ -22,9 +21,9 @@ class HttpRequestHeaderTest {
 
         Map<HttpHeaders, String> result = httpRequestHeader.getHeaders();
         assertThat(result).contains(
-                Map.entry(HttpHeaders.HOST,"www.nowhere123.com"),
-                Map.entry(HttpHeaders.ACCEPT,"image/gif, image/jpeg, */*"),
-                Map.entry(HttpHeaders.ACCEPT_LANGUAGE,"en-us")
+                Map.entry(HttpHeaders.HOST, "www.nowhere123.com"),
+                Map.entry(HttpHeaders.ACCEPT, "image/gif, image/jpeg, */*"),
+                Map.entry(HttpHeaders.ACCEPT_LANGUAGE, "en-us")
         );
     }
