+++ src/test/java/webserver/http/PathTest.java
@@ -1,4 +1,4 @@
-package model;
+package webserver.http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
